Bug Fix - Issue 240 - JXcore dead locks when expiring items out of the memory store
When expiring items out of the memory store, the locks are taken in the wrong order.
The modifier functions in src/wrappers/memory_wrap.cc take the locks in the reverse
order to this function, hence causes a race condition which can end in two threads
holding and waiting for a lock that the other holds.

Modified this function so that the timers lock is taken first, collects a list
of keys to delete, releases the timer lock and takes the store lock. The
result being that a dead lock is impossible.

Tested on a machine running jx core under high load. Dead lock has now not occured for
72 hours. Previously dead locks occured every 6 hours.