rewrite of store.shared.safeBlock
The old version of safeBlock required the block function to be the same for
all attempts to modify the key. If a closure was used as a block
unexpected results occured. In fact the closure used would be the the
last closure used against the call to safeBlock, however this closure
would be called multiple times.

So it would work if the closure was simply incrementing the key, other
modifications that depend on values of the closure would produce
unexpected results.

The following code demonstrates the issue. On the old code the you will
find that the only one key out of obj are cleared per thread, because
the allDone closure depend of r. Only the last key is cleared, the
old keys are left.

The new code works as expected.

var method = {
	define: function () {
		var obj = {};
		var objCount = 0;

	},
  logic: function(r) {
		var allDone = function() {
			if(obj[r]) {
				delete obj[r];
				objCount--;
			}
			process.release();
			process.sendToMain({count: objCount});
		};

		obj[r] = 1;
		objCount++;
		jxcore.store.shared.safeBlock("myNumber", function () {
			allDone();
		});
		process.keepAlive();
	}
};

var counts = {};
jxcore.tasks.on('message', function (threadId, param) {
	counts[threadId] = param.count;
});

for(var i=0; i<1000; i++)  {
	jxcore.tasks.addTask(method, i );
}

jxcore.tasks.on('emptyQueue', function () {
	for(var threadId in counts) {
		console.log(threadId + ': ' + counts[threadId]);
	}
});