Efficiently repair reading of signed files
This change makes use of garbage bytes (16) at the end of the file as
information to locate the required data to load for a native executable
efficiently. Additionally, it contains an efficient algorithm to look
for a the signature at the end of a signed file, if necessary, and
skipping it directly to the intended last byte of the file. The maximum
data size and custom offset using this algorithm are 4GB and the maximum
signature size at the end of the file is 10KB.