Rewrite for better fix
Found a better solution by re-using the code that was originally
supposed to store the size of the appBuffer -- it now stores the exact
physical location of the new indexing data which holds the size of the
appBuffer. This method now allows the loading data to be found
efficiently regardless of the data at the end of the file and also
prevents the loading process from being tampered with without
invalidating the digital signature.