omit space in version string
prior art:

``` sh
$ node -v
v0.10.40
$ npm -v
1.4.14
```

Personally, I would be okay with dropping the `v` entirely, but at the very least, removing the space would be beneficial. Aside from not following convention, the version string seems to be used as the lockfile name for npm. Which means jx is writing a file with a space in the name; that's just not nice. :)