jxcore: moved node::removeCommons() outside AutoScope scope
When the AutoScope object goes out of scope in the JXEngine::Destroy method it uses members of main_node_ that prior to this fix had already been destroyed by the node::removeCommons() call. This fix moves the node::removeCommons autoside the lifetime scope of the AutoScope object, such that it's guaranteed that all members of main_node_ are still dereferenceable.