jxcore: assign correct value to 'hasStdFds'
Fix #608 somehow introduced a broken fix for supporting embedded headless client (like e.g. Windows UI applications). The 'hasStdFds' should indicate whether stdout/stdin/stderr is available, in #608 however it was true when the handle type for stdout was unknown. Adjusting this to do the opposite caused a crash in a sub engine being started on a different thread. This is due to the loopId member variable not being initialized. Assigning 'threadId' to the 'loopId' (like the main thread implicitly does) seems to fix this problem.